refactor: extract auth utils and fix login navigation
