+++ App/config/firebaseApp.ts
@@ -1,4 +1,4 @@
 // Firebase API helper utilities
 export const API_URL = process.env.EXPO_PUBLIC_API_URL || '';
 
-export { getAuthHeader, getAuthHeaders } from '@/utils/TokenManager';
+export { getAuthHeader, getAuthHeaders } from '@/utils/authUtils';
