+++ App/utils/TokenManager.ts
@@ -1,6 +1,5 @@
 import { getDocument, setDocument } from '@/services/firestoreService';
 import { ensureAuth } from '@/utils/authGuard';
-import { getIdToken as fbGetIdToken, getCurrentUserId as fbGetCurrentUserId } from '@/lib/auth';
 
 export const getTokenCount = async () => {
   const uid = await ensureAuth();
@@ -64,28 +63,7 @@ export const syncSubscriptionStatus = async () => {
   }
 };
 
-export async function getIdToken(forceRefresh = false) {
-  return fbGetIdToken(forceRefresh);
-}
-
-export async function getCurrentUserId(): Promise<string | null> {
-  return fbGetCurrentUserId();
-}
-
-export async function getAuthHeader() {
-  const token = await getIdToken(true);
-  if (!token) throw new Error('Unable to refresh ID token');
-  return { Authorization: `Bearer ${token}` };
-}
-
-export async function getAuthHeaders() {
-  const { Authorization } = await getAuthHeader();
-  return { Authorization, 'Content-Type': 'application/json' };
-}
-
-export async function getToken(forceRefresh = false) {
-  return getIdToken(forceRefresh);
-}
+export { getIdToken, getCurrentUserId, getAuthHeader, getAuthHeaders, getToken } from './authUtils';
 
 export function init() {
   console.log('✅ TokenManager initialized');
