+++ App/utils/gusBugLogger.ts
@@ -1,5 +1,5 @@
 import { signOutAndRetry, logTokenIssue } from '@/services/authService';
-import { getIdToken } from '@/utils/TokenManager';
+import { getIdToken } from '@/utils/authUtils';
 import { showPermissionDenied } from '@/utils/gracefulError';
 import { useAuthStore } from '@/state/authStore';
